moved Header.jsx to components folder and removed moved SongControl from App.jsx
